Added exporting feature, and basic requirement

+++ core/api.py
@@ -6,14 +6,18 @@ from rest_framework.response import Response
 from rest_framework.viewsets import ViewSet
 from rest_framework import status
 from rest_framework.decorators import action
-from rest_framework.exceptions import MethodNotAllowed
+from rest_framework.exceptions import MethodNotAllowed, PermissionDenied
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAdminUser
+import random
 
 from core.forms import (
     AddCourseForm,
+    UpdateCourseForm,
     AddDepartmentForm,
     AddVenueCategoryForm,
     AddVenueForm,
     AddStaffForm,
+    UpdateStaffForm,
 )
 
 from core.models import Course, Department, Staff, Venue, VenueCategory
@@ -27,16 +31,21 @@ from core.serializers import (
     VenueSerializer,
 )
 
-
-@api_view()
-@csrf_exempt
-def courses(request: Request):
-    _courses = Course.objects.all()
-    return Response(CourseSerializer(_courses, many=True).data)
+from account.permissions import (
+    IsAdmin,
+    IsStaff,
+    IsStudent,
+    userIsAdmin,
+    userIsStaff,
+    userIsStudent,
+)
+from rest_framework.permissions import IsAuthenticated
 
 
 class CourseViewSet(ViewSet):
     def list(self, request: Request):
+        """Returns the list off all the courses in the system"""
+
         _courses = Course.objects.all()
         queries = request.query_params
         query = queries.get("q", None)
@@ -47,6 +56,18 @@ class CourseViewSet(ViewSet):
         return Response(CourseSerializer(_courses, many=True).data)
 
     def create(self, request: Request):
+        """Accept a valid form to create a new course
+
+        Args:
+            request (Request): _description_
+
+        Returns:
+            _type_: _description_
+        """
+
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin user can create new courses")
+
         form = AddCourseForm(request.data)  # type: ignore
 
         if not form.is_valid():
@@ -61,7 +82,64 @@ class CourseViewSet(ViewSet):
             status=status.HTTP_201_CREATED, data=CourseSerializer(course).data
         )
 
-    @action(detail=False, methods=("POST", "DELETE"))
+    def update(self, request: Request, pk: None):
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin users can update courses")
+
+        try:
+            course = Course.objects.get(pk=pk)
+
+            form = UpdateCourseForm(request.data)  # type: ignore
+
+            if not form.is_valid():
+                return Response(
+                    status=status.HTTP_400_BAD_REQUEST,
+                    data={"details": "Invalid form provided", "errors": form.errors},
+                )
+
+            data = form.cleaned_data
+
+            used_code = (
+                Course.objects.filter(code=data["code"]).exclude(pk=pk).count() > 0
+            )
+
+            if used_code:
+                return Response(
+                    status=status.HTTP_409_CONFLICT,
+                    data={"details": "Code been used by another course"},
+                )
+
+            course.title = data["title"]
+            course.code = data["code"]
+            course.department = data["department"]
+            course.departments.set(data["departments"])
+            course.student_count = data["student_count"]
+            course.level = data["level"]
+            course.semester = data["semester"]
+            course.save()
+
+            return Response(CourseSerializer(course).data)
+
+        except Course.DoesNotExist:
+            return Response(
+                status=status.HTTP_404_NOT_FOUND, data={"details": "No course found"}
+            )
+
+    def retrieve(self, request: Request, pk=None):
+        try:
+            course = Course.objects.get(pk=pk)
+            return Response(CourseSerializer(course).data)
+        except Course.DoesNotExist:
+            return Response(
+                status=status.HTTP_404_NOT_FOUND,
+                data={"details": "Course does not exist"},
+            )
+
+    @action(
+        detail=False,
+        methods=("POST", "DELETE"),
+        permission_classes=(IsAuthenticated, IsAdmin),
+    )
     def multiple(self, request: Request):
         request_method = request._request.method
 
@@ -77,7 +155,9 @@ class CourseViewSet(ViewSet):
         else:
             raise MethodNotAllowed(method=request_method)
 
-    @action(methods=("DELETE",), detail=False)
+    @action(
+        methods=("DELETE",), detail=False, permission_classes=(IsAuthenticated, IsAdmin)
+    )
     def multiple_delete(self, request: Request):
         ids = request.data
 
@@ -101,9 +181,114 @@ class CourseViewSet(ViewSet):
 
         return Response(CourseSerializer(course, many=True).data)
 
+    @action(detail=False, permission_classes=(IsAuthenticated, IsAdmin))
+    def generate_random_courses(self, request: Request):
+        levels = [100, 200, 300, 400]
+        try:
+            count: int = int(request.query_params.get("count", 0))
+            department_max: int = int(request.query_params.get("department_max", 0))
+            level: int = int(request.query_params.get("level", 0))
+            semester: int = int(request.query_params.get("semester", 1))
+
+            if count < 1 or department_max < 0:
+                return Response(
+                    status=status.HTTP_400_BAD_REQUEST,
+                    data={
+                        "detail": "count and department max needs to be greather than 0"
+                    },
+                )
+            if level not in levels:
+                return Response(
+                    status=status.HTTP_400_BAD_REQUEST,
+                    data={"detail": "Invalid level"},
+                )
+
+            departments = set(Department.objects.all())
+            department_count = len(departments)
+            department_courses_count = {}
+            completed_departments = set([])
+
+            courses_added = []
+
+            for current_index in range(count):
+                pool = departments.difference(completed_departments)
+                print(f"Index : {current_index}, Pool Size : {len(pool)}")
+
+                if len(pool) == 0:
+                    break
+
+                else:
+                    department = random.choice(list(pool))
+
+                    if department not in department_courses_count:
+                        department_courses_count[department] = 0
+
+                    course = Course()
+                    course.title = f"Course {current_index + 1}"
+                    course.code = (
+                        f"{department.code}{random.randrange(level+1, level+99)}"
+                    )
+                    course.department = department
+                    course.level = level
+                    course.semester = semester
+                    course.student_count = random.randint(30, 300)
+                    selected_departments = set()
+
+                    if course.code.endswith("101") or random.randint(1, 30) % 8 == 0:
+                        # Is a shared course
+                        shared_department_count = random.randint(1, department_count)
+                        selected_departments = set(
+                            random.choices(list(departments), k=shared_department_count)
+                        )
+                        selected_departments = selected_departments.difference(
+                            completed_departments
+                        )
+
+                    selected_departments.add(department)
+                    try:
+                        course.save()
+                        course.departments.set(selected_departments)
+                        course.save()
+                        department_courses_count[department] = (
+                            department_courses_count[department] + 1
+                        )
+                        courses_added.append(course)
+
+                        if department_courses_count[department] >= department_max:
+                            completed_departments.add(department)
+                    except Exception:
+                        pass
+
+            return Response(
+                data={
+                    "detail": f"Added {len(courses_added)} courses",
+                    "count": count,
+                    "courses": PlainCourseSerializer(courses_added, many=True).data,
+                }
+            )
+
+        except TypeError as e:
+            return Response(
+                status=status.HTTP_400_BAD_REQUEST,
+                data={"detail": "wrong, missing or invalid count"},
+            )
+        except ValueError as e:
+            return Response(
+                status=status.HTTP_400_BAD_REQUEST,
+                data={"detail": "missing or invalid count", "extra": str(e)},
+            )
+
 
 class DepartmentViewSet(ViewSet):
     def list(self, request: Request):
+        """Returns list of departments in the system
+
+        Args:
+            request (Request): _description_
+
+        Returns:
+            _type_: _description_
+        """
         departments = Department.objects.all()
         queries = request.query_params
         query = queries.get("q", None)
@@ -114,6 +299,18 @@ class DepartmentViewSet(ViewSet):
         return Response(DepartmentSerializer(departments, many=True).data)
 
     def create(self, request: Request):
+        """Accept valid from to create a new department into the system
+
+        Args:
+            request (Request): _description_
+
+        Returns:
+            _type_: _description_
+        """
+
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin user can create new department")
+
         form = AddDepartmentForm(request.data)  # type: ignore
         if not form.is_valid():
             return Response(
@@ -135,7 +332,7 @@ class DepartmentViewSet(ViewSet):
             courses = Course.objects.filter(department=department)
             return Response(
                 {
-                    **DepartmentSerializer(department).data,
+                    **DepartmentSerializer(department).data,  # type: ignore
                     "staffs": PlainStaffSerializer(staffs, many=True).data,
                     "courses": PlainCourseSerializer(courses, many=True).data,
                 }
@@ -145,7 +342,11 @@ class DepartmentViewSet(ViewSet):
                 {"Details": "Department not found"}, status=status.HTTP_404_NOT_FOUND
             )
 
-    @action(detail=False, methods=("POST", "DELETE"))
+    @action(
+        detail=False,
+        methods=("POST", "DELETE"),
+        permission_classes=(IsAuthenticated, IsAdmin),
+    )
     def multiple(self, request: Request):
         request_method = request._request.method
 
@@ -161,7 +362,9 @@ class DepartmentViewSet(ViewSet):
         else:
             raise MethodNotAllowed(method=request_method)
 
-    @action(methods=("DELETE",), detail=False)
+    @action(
+        methods=("DELETE",), detail=False, permission_classes=(IsAuthenticated, IsAdmin)
+    )
     def multiple_delete(self, request: Request):
         ids = request.data
 
@@ -178,7 +381,8 @@ class DepartmentViewSet(ViewSet):
         return Response(DepartmentSerializer(departments, many=True).data)
 
     def delete(self, request: Request, pk=None):
-        print("deleting")
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin user can delete a department")
         try:
             department = Department.objects.get(pk=pk)
             department.delete()
@@ -197,30 +401,58 @@ class DepartmentViewSet(ViewSet):
         return Response(DepartmentSerializer(departments, many=True).data)
 
     @action(detail=True)
-    def courses(self, request: Request, pk = None):
+    def courses(self, request: Request, pk=None):
+        """Returns list of courses that belongs to the department with id of {pk}
+
+        Args:
+            request (Request): _description_
+            pk (_type_, optional): _description_. Defaults to None.
+
+        Returns:
+            _type_: _description_
+        """
         try:
             department = Department.objects.get(pk=pk)
-            courses = Course.objects.filter(department = department)
-            return Response(CourseSerializer(courses, many = True).data)
-        except Department.DoesNotExist:
-            return Response(
-                status=status.HTTP_404_NOT_FOUND,
-                data={"details": "Department not found"},
-            )
-    @action(detail=True)
-    def staffs(self, request: Request, pk = None):
-        try:
-            department = Department.objects.get(pk=pk)
-            staffs = Staff.objects.filter(department = department)
-            return Response(StaffSerializer(staffs, many = True).data)
+            courses = Course.objects.filter(department=department)
+            return Response(CourseSerializer(courses, many=True).data)
         except Department.DoesNotExist:
             return Response(
                 status=status.HTTP_404_NOT_FOUND,
                 data={"details": "Department not found"},
             )
 
+    @action(detail=True)
+    def staffs(self, request: Request, pk=None):
+        """Returns list of staffs that belong to th department with id of {pk}
+
+        Args:
+            request (Request): _description_
+            pk (_type_, optional): _description_. Defaults to None.
+
+        Returns:
+            _type_: _description_
+        """
+        try:
+            department = Department.objects.get(pk=pk)
+            staffs = Staff.objects.filter(department=department)
+            return Response(StaffSerializer(staffs, many=True).data)
+        except Department.DoesNotExist:
+            return Response(
+                status=status.HTTP_404_NOT_FOUND,
+                data={"details": "Department not found"},
+            )
+
+
 class VenueViewSet(ViewSet):
     def list(self, request: Request):
+        """Returns he list of venues in the system
+
+        Args:
+            request (Request): _description_
+
+        Returns:
+            _type_: _description_
+        """
         venues = Venue.objects.all()
         query = request.query_params.get("q", None)
         if query:
@@ -230,6 +462,22 @@ class VenueViewSet(ViewSet):
 
     @action(methods=("GET", "POST", "DELETE"), detail=False)
     def categories(self, request: Request):
+        """Handles listing, deleting and adding new venue category based on the method type
+        post : for creating
+        delete: for deleting
+        get: for listing
+
+        Args:
+            request (Request): _description_
+
+        Raises:
+            PermissionDenied: _description_
+
+        Returns:
+            _type_: _description_
+        """
+        if not userIsAdmin(request) and request._request.method in ("DELETE", "POST"):
+            raise PermissionDenied("This action is only available to admin users")
         if request._request.method == "DELETE":
             # Handles for deleting venue categories
             categories = VenueCategory.objects.filter(Q(pk__in=request.data))
@@ -268,6 +516,9 @@ class VenueViewSet(ViewSet):
         """
         Handles creation of new venue
         """
+
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin user can create new venues")
         form = AddVenueForm(request.data)  # type: ignore
 
         if not form.is_valid():
@@ -296,7 +547,9 @@ class VenueViewSet(ViewSet):
         else:
             raise MethodNotAllowed(method=request_method)
 
-    @action(methods=("DELETE",), detail=False)
+    @action(
+        methods=("DELETE",), detail=False, permission_classes=(IsAuthenticated, IsAdmin)
+    )
     def multiple_delete(self, request: Request):
         """
         Handles deletion of multiple venues at a time
@@ -327,6 +580,8 @@ class StafftViewSet(ViewSet):
         return Response(StaffSerializer(staffs, many=True).data)
 
     def create(self, request: Request):
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin users can add new staff")
         form = AddStaffForm(request.data)  # type: ignore
         if not form.is_valid():
             return Response(
@@ -341,6 +596,48 @@ class StafftViewSet(ViewSet):
         staff = form.save()
         return Response(StaffSerializer(staff).data)
 
+    def update(self, request: Request, pk: None):
+        try:
+            staff = Staff.objects.get(pk=pk)
+
+            if not userIsAdmin(request):
+                raise PermissionDenied("Only admin users can add new staff")
+            form = UpdateStaffForm(request.data)  # type: ignore
+            if not form.is_valid():
+                return Response(
+                    status=status.HTTP_400_BAD_REQUEST,
+                    data={
+                        "details": "Invalid form",
+                        "errors": form.errors,
+                        "fields": form.errors.keys(),
+                    },
+                )
+
+            _staff = form.cleaned_data
+            staff.name = _staff["name"]
+            staff.can_invigilate = _staff["can_invigilate"]
+            staff.can_supervise = _staff["can_supervise"]
+            staff.staff_id = _staff["staff_id"]
+            staff.department = _staff["department"]
+
+            if (
+                Staff.objects.filter(staff_id=_staff["staff_id"])
+                .exclude(pk=staff.pk)
+                .count()
+                > 0
+            ):
+                return Response(
+                    status=status.HTTP_409_CONFLICT,
+                    data={"details": "Staff Id used by another staff"},
+                )
+
+            staff.save()
+            return Response(StaffSerializer(staff).data)
+        except Staff.DoesNotExist:
+            return Response(
+                status=status.HTTP_404_NOT_FOUND, data={"details": "No staff found"}
+            )
+
     def retrieve(self, request: Request, pk=None):
         try:
             staff = Staff.objects.get(pk=pk)
@@ -350,7 +647,11 @@ class StafftViewSet(ViewSet):
                 status=status.HTTP_404_NOT_FOUND, data={"details": "Staff not found"}
             )
 
-    @action(detail=False, methods=("POST", "DELETE"))
+    @action(
+        detail=False,
+        methods=("POST", "DELETE"),
+        permission_classes=(IsAuthenticated, IsAdmin),
+    )
     def multiple(self, request: Request):
         request_method = request._request.method
 
@@ -359,7 +660,7 @@ class StafftViewSet(ViewSet):
         if request_method == "POST":
             staffs = Staff.objects.filter(Q(pk__in=staffs_pks))
             return Response(StaffSerializer(staffs, many=True).data)
-        elif request_method == "DETETE":
+        elif request_method == "DELETE":
             staffs = Staff.objects.filter(Q(pk__in=staffs_pks))
             staffs.delete()
             return Response(StaffSerializer(staffs, many=True).data)
@@ -367,6 +668,8 @@ class StafftViewSet(ViewSet):
             raise MethodNotAllowed(method=request_method)
 
     def delete(self, request: Request, pk=None):
+        if not userIsAdmin(request):
+            raise PermissionDenied("Only admin user can delete a staff")
         try:
             staff = Staff.objects.get(pk=pk)
             staff.delete()
